+++ src/main/java/com/example/jpaschedule/config/WebConfig.java
@@ -1,10 +1,12 @@
 package com.example.jpaschedule.config;
 
 import com.example.jpaschedule.filter.LoginFilter;
+import com.example.jpaschedule.filter.interceptor.SessionInterceptor;
 import jakarta.servlet.Filter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -17,5 +19,12 @@ public class WebConfig implements WebMvcConfigurer {
         filterRegistrationBean.setOrder(1);
         filterRegistrationBean.addUrlPatterns("/*");
         return filterRegistrationBean;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new SessionInterceptor())
+                .addPathPatterns("/**")
+                .excludePathPatterns("/api/v1/members/signup", "/api/v1/login");
     }
 }
