+++ src/main/java/com/example/jpaschedule/controller/LoginController.java
@@ -4,8 +4,6 @@ import com.example.jpaschedule.common.Const;
 import com.example.jpaschedule.dto.request.LoginRequestDto;
 import com.example.jpaschedule.dto.response.MemberResponseDto;
 import com.example.jpaschedule.service.LoginService;
-import com.example.jpaschedule.service.MemberService;
-import com.example.jpaschedule.service.context.MemberContext;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +26,7 @@ public class LoginController {
     public ResponseEntity<Void> login(@RequestBody LoginRequestDto dto, HttpServletRequest request) {
         MemberResponseDto responseDto = loginService.login(dto.getEmail(), dto.getPassword());
 
-        HttpSession session =  request.getSession();
+        HttpSession session = request.getSession();
 
         session.setAttribute(Const.SESSION_KEY, responseDto.getId());
         return ResponseEntity.ok().build();
