+++ src/main/java/com/example/jpaschedule/filter/LoginFilter.java
@@ -1,7 +1,7 @@
 package com.example.jpaschedule.filter;
 
 import com.example.jpaschedule.common.Const;
-import com.example.jpaschedule.service.context.MemberContext;
+import com.example.jpaschedule.filter.context.MemberContext;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -14,14 +14,12 @@ import java.io.IOException;
 @Slf4j
 public class LoginFilter implements Filter {
 
-    private static final String[] WHITE_LIST = new String[]{"/", "/api/v1/members/signup", "/api/v1/login", "/api/v1/logout"};
+    private static final String[] WHITE_LIST = new String[]{"/api/v1/members/signup", "/api/v1/login", "/api/v1/logout"};
 
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
         HttpServletRequest httpRequest = (HttpServletRequest) request;
         String requestURI = httpRequest.getRequestURI();
-
-        log.info("login filter");
 
         // TODO 사용 여부 확인 후 삭제
         HttpServletResponse httpResponse = (HttpServletResponse) response;
@@ -34,7 +32,7 @@ public class LoginFilter implements Filter {
                 throw new RuntimeException("로그인이 필요한 요청입니다.");
             }
             MemberContext.setMemberId((Long) session.getAttribute(Const.SESSION_KEY));
-            log.info("Login successful");
+            log.info("로그인 성공");
         }
         chain.doFilter(request, response);
     }
