+++ src/main/java/com/example/jpaschedule/filter/interceptor/SessionInterceptor.java
@@ -1,0 +1,19 @@
+package com.example.jpaschedule.filter.interceptor;
+
+import com.example.jpaschedule.filter.context.MemberContext;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+@Slf4j
+public class SessionInterceptor implements HandlerInterceptor {
+
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+        HandlerInterceptor.super.afterCompletion(request, response, handler, ex);
+
+        log.info("afterCompletion memberContext: {}", MemberContext.getMemberId());
+        MemberContext.clear();
+    }
+}
