+++ src/main/java/com/example/jpaschedule/service/ScheduleService.java
@@ -6,7 +6,7 @@ import com.example.jpaschedule.entity.Member;
 import com.example.jpaschedule.entity.Schedule;
 import com.example.jpaschedule.repository.MemberRepository;
 import com.example.jpaschedule.repository.ScheduleRepository;
-import com.example.jpaschedule.service.context.MemberContext;
+import com.example.jpaschedule.filter.context.MemberContext;
 import com.example.jpaschedule.util.EmptyTool;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -33,7 +33,6 @@ public class ScheduleService {
     }
 
     public List<ScheduleResponseDto> findAll() {
-        // 로그인한 유저의 일정 전체 조회
         List<Schedule> schedules = scheduleRepository.findAllByMember_Id(MemberContext.getMemberId());
         List<ScheduleResponseDto> scheduleResponseDtos = new ArrayList<>();
         for (Schedule schedule : schedules) {
